// src/lib/notifications/types.ts
export interface FCMNotificationItem {
  send: (...data: unknown[]) => void
}

export enum NotificationTypes {
  NEW_BID_ON_AUCTION = 'NEW_BID_ON_AUCTION',
  AUCTION_UPDATED = 'AUCTION_UPDATED',
  BID_REMOVED_ON_AUCTION = 'BID_REMOVED_ON_AUCTION',
  BID_ACCEPTED_ON_AUCTION = 'BID_ACCEPTED_ON_AUCTION',
  BID_REJECTED_ON_AUCTION = 'BID_REJECTED_ON_AUCTION',
  REVIEW_RECEIVED = 'REVIEW_RECEIVED',
  NEW_MESSAGE = 'NEW_MESSAGE',
  SYSTEM = 'SYSTEM',
  SOMEONE_ELSE_ADDED_BID_TO_SAME_AUCTION = 'SOMEONE_ELSE_ADDED_BID_TO_SAME_AUCTION',
  BID_WAS_SEEN = 'BID_WAS_SEEN',
  NEW_FOLLOWER = 'NEW_FOLLOWER',
  AUCTION_FROM_FAVOURITES_HAS_BID = 'AUCTION_FROM_FAVOURITES_HAS_BID',
  NEW_AUCTION_FROM_FOLLOWING = 'NEW_AUCTION_FROM_FOLLOWING',
  AUCTION_ADDED_TO_FAVOURITES = 'AUCTION_ADDED_TO_FAVOURITES',
  FAVOURITE_AUCTION_PRICE_CHANGE = 'FAVOURITE_AUCTION_PRICE_CHANGE',
  MY_AUCTION_STARTED = 'MY_AUCTION_STARTED',
  AUCTION_FROM_FAVOURITES_STARTED = 'AUCTION_FROM_FAVOURITES_STARTED',
  ACCOUNT_VERIFIED = 'ACCOUNT_VERIFIED',
  NEW_COMMENT_ON_AUCTION = 'NEW_COMMENT_ON_AUCTION',
  COMMENT_ON_SAME_AUCTION = 'COMMENT_ON_SAME_AUCTION',
  REPLY_ON_AUCTION_COMMENT = 'REPLY_ON_AUCTION_COMMENT',
  NEW_OFFER_ON_AUCTION = 'NEW_OFFER_ON_AUCTION',
}